Move copyRealmSchema into Observation model
